pyproject.toml
[project]
name = "octopusos"
version = "3.2.0"
description = "System-level, project-agnostic AI Agent orchestration system"
requires-python = ">=3.13"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "click>=8.1.7",
    "openai>=1.58.1",
    "jinja2>=3.1.5",
    "jsonschema>=4.23.0",
    "rich>=13.9.4",
    "croniter>=1.4.1",
    "networkx>=3.1",
    "pyyaml>=6.0",
    "anthropic>=0.18.0",
    "docker>=6.1.0",
    "gitpython>=3.1.46",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "psutil>=5.9.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "sentry-sdk[fastapi]>=2.50.0",
    "cryptography>=42.0.0",
    "python-ulid>=2.2.0",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.1.0",  # Session signing for CSRF protection (Task #36)
    "duckduckgo-search>=6.3.11",  # Web search functionality (DuckDuckGo)
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "webrtcvad>=2.0.10",  # Voice Activity Detection (VAD)
    "numpy>=1.24.0",  # Required for audio processing
    "PyNaCl>=1.5.0",  # Ed25519 signature verification for Discord
    "grpcio>=1.60.0",  # gRPC for Voice Worker Sidecar (v0.2)
    "grpcio-tools>=1.60.0",  # gRPC protocol compiler for Voice Worker (v0.2)
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.5",
]
voice = [
    "faster-whisper>=1.0.0; python_version < '3.14'",
]
vector = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
]
postgres = [
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
octopusos = "octopusos.cli.main:cli"
memoryos = "memoryos.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ============================================================
# Hatchling build configuration (CRITICAL for wheel packaging)
# ============================================================
[tool.hatch.build]
# Include all non-.py files within packages
# This is the KEY setting for static assets to be included in wheel
include = [
    "os/octopusos/**/*.yaml",
    "os/octopusos/**/*.yml",
    "os/octopusos/**/*.json",
    "os/octopusos/**/*.sql",
    "os/octopusos/**/*.html",
    "os/octopusos/**/*.css",
    "os/octopusos/**/*.js",
    "os/octopusos/**/*.png",
    "os/octopusos/**/*.jpg",
    "os/octopusos/**/*.svg",
    "os/octopusos/**/*.ico",
    "os/octopusos/**/*.woff",
    "os/octopusos/**/*.woff2",
    "os/octopusos/**/*.ttf",
    "os/octopusos/**/*.eot",
    "os/octopusos/**/*.map",
    "os/octopusos/**/*.min.js",
    "os/octopusos/**/*.min.css",
    "os/octopusos/**/*.proto",  # gRPC protocol files (v0.2)
    "os/memoryos/**/*.yaml",
    "os/memoryos/**/*.yml",
    "os/memoryos/**/*.json",
]

[tool.hatch.build.targets.wheel]
# Explicitly list packages (auto-discovery should work, but be explicit)
packages = ["os/octopusos", "os/memoryos"]

# CRITICAL: Force-include directories that might not be auto-detected
# Hatchling force-include syntax: "source_path" = "destination_in_wheel"
force-include = { "rules" = "rules" }

[tool.hatch.build.targets.sdist]
# For sdist, include everything (also respects MANIFEST.in)
include = [
    "/os/octopusos",
    "/os/memoryos",
    "/rules",
    "/examples",
    "/scripts/gates",
    "/scripts/demos",
    "/MANIFEST.in",
    "/README.md",
    "/LICENSE",
    "/NOTICE.md",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-v",
    "-m",
    "not integration and not slow",
]
asyncio_mode = "auto"
markers = [
    "asyncio: Async test functions managed by pytest-asyncio",
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests (> 1s)",
    "network: Tests requiring network access",
]

[tool.coverage.run]
source = ["os/octopusos"]
omit = [
    "tests/*",
    "*/tests/*",
    "**/test_*.py",
    "*/migrations/*",
    "**/migrations/*",
    "*/conftest.py",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/.venv/*",
    "*/venv/*",
    "os/octopusos/webui/static/*",
    "os/octopusos/webui/templates/*",
]
branch = true
parallel = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
    "pass",
    "\\.\\.\\.",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"
